change pitchBend variable to int

+++ src/ios/CDVInstrument.m
@@ -165,7 +165,7 @@
 // noteOn(noteNumber, velocity)
 - (void)pitchBend:(CDVInvokedUrlCommand*)command {
     if (hasSoundFont) {
-        Byte pitchBend = [[command.arguments objectAtIndex:0] intValue];
+        int pitchBend = [[command.arguments objectAtIndex:0] intValue];
         Byte channel = [[command.arguments objectAtIndex:1] intValue];
         NSLog(@"PitchBend %i", pitchBend);
         [self.instrument sendPitchBend:pitchBend onChannel:channel];
